fix(app): one canonical readinessScore for soma's model (#359)
soma's model composite was rendered three ways: overview hero
Math.round(min(1,z)*100), comparison chart normal-percentile, /training raw z.
The overview mapping is broken for normal days (z=0 -> 0, z<0 -> negative).

Add a shared readinessScore(z) helper (bounded normal-distribution percentile,
0-100) and use it in the overview hero and the comparison chart, so the same z
reads the same everywhere. z=0->50, z=-2->2, z=2->98.

Web surfaces + VO2max/HRV reconciliation tracked separately.

Closes #359
